requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
facenet-pytorch==2.5.2
filelock==3.7.1
huggingface-hub==0.0.8
idna==3.3
joblib==1.1.0
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.20.0
pyparsing==3.0.9
regex==2022.7.9
requests==2.28.1
sacremoses==0.0.53
scikit-learn==1.1.1
scipy==1.8.1
sentencepiece==0.1.96
six==1.16.0
tabulate==0.8.10
threadpoolctl==3.1.0
timm==0.4.12
tokenizers==0.9.2
torch==1.8.1+cu101
torchaudio==0.8.1
torchvision==0.9.1+cu101
tqdm==4.64.0
transformers==3.4.0
typing-extensions==4.3.0
urllib3==1.26.10
